refactor(data handler function separated): data handler functions separated
